[date]/route.tsを[id]に統括
